Setup MySQL at runtime

+++ install_mysql.sh
@@ -1,22 +1,26 @@
 #/bin/sh
 
-# Setup MySQL data dir
-mysql_install_db --user=mysql --datadir="${DB_DATA_PATH}"
+if [[ ! -d ${DB_DATA_PATH}/mysql ]]; then
 
-# Set MySQL root password
-mysqld_safe --datadir="${DB_DATA_PATH}" &
-sleep 5 # Wait for startup
-mysqladmin -u root password "${DB_ROOT_PASS}"
+    # Setup MySQL data dir
+    mysql_install_db --user=mysql --datadir="${DB_DATA_PATH}"
 
-# Create user for ampache
-echo "GRANT ALL ON *.* TO ${DB_USER}@'127.0.0.1' IDENTIFIED BY '${DB_PASS}' WITH GRANT OPTION;" > /tmp/sql
-echo "GRANT ALL ON *.* TO ${DB_USER}@'localhost' IDENTIFIED BY '${DB_PASS}' WITH GRANT OPTION;" >> /tmp/sql
-echo "GRANT ALL ON *.* TO ${DB_USER}@'::1' IDENTIFIED BY '${DB_PASS}' WITH GRANT OPTION;" >> /tmp/sql
-echo "DELETE FROM mysql.user WHERE User='';" >> /tmp/sql
-echo "DROP DATABASE test;" >> /tmp/sql
-echo "FLUSH PRIVILEGES;" >> /tmp/sql
-cat /tmp/sql | mysql -u root --password="${DB_ROOT_PASS}"
+    # Set MySQL root password
+    mysqld_safe --datadir="${DB_DATA_PATH}" &
+    sleep 5 # Wait for startup
+    mysqladmin -u root password "${DB_ROOT_PASS}"
 
-# MySQL config
-sed -i "s|max_allowed_packet\s*=\s*1M|max_allowed_packet = ${MAX_ALLOWED_PACKET}|g" /etc/mysql/my.cnf
-sed -i "s|max_allowed_packet\s*=\s*16M|max_allowed_packet = ${MAX_ALLOWED_PACKET}|g" /etc/mysql/my.cnf
+    # Create user for ampache
+    echo "GRANT ALL ON *.* TO ${DB_USER}@'127.0.0.1' IDENTIFIED BY '${DB_PASS}' WITH GRANT OPTION;" > /tmp/sql
+    echo "GRANT ALL ON *.* TO ${DB_USER}@'localhost' IDENTIFIED BY '${DB_PASS}' WITH GRANT OPTION;" >> /tmp/sql
+    echo "GRANT ALL ON *.* TO ${DB_USER}@'::1' IDENTIFIED BY '${DB_PASS}' WITH GRANT OPTION;" >> /tmp/sql
+    echo "DELETE FROM mysql.user WHERE User='';" >> /tmp/sql
+    echo "DROP DATABASE test;" >> /tmp/sql
+    echo "FLUSH PRIVILEGES;" >> /tmp/sql
+    cat /tmp/sql | mysql -u root --password="${DB_ROOT_PASS}"
+
+    # MySQL config
+    sed -i "s|max_allowed_packet\s*=\s*1M|max_allowed_packet = ${MAX_ALLOWED_PACKET}|g" /etc/mysql/my.cnf
+    sed -i "s|max_allowed_packet\s*=\s*16M|max_allowed_packet = ${MAX_ALLOWED_PACKET}|g" /etc/mysql/my.cnf
+
+fi
